feat(sbt): :hammer: refactor project structure
