+++ play-actuator-core/src/main/scala/play/actuator/ActuatorEnum.scala
