+++ play-actuator-core/src/main/scala/play/actuator/health/indicator/BaseHealthIndicator.scala
